@@ -3,20 +3,13 @@ package br.uff.tecnomarias.domain.entity;
 import br.uff.tecnomarias.domain.enums.TipoPessoa;
 
 import javax.persistence.*;
-import javax.transaction.Transactional;
-import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 import java.util.function.Consumer;
 
 @Entity
 @Inheritance(strategy = InheritanceType.JOINED)
-@DiscriminatorColumn(name="TIPO_PESSOA", discriminatorType = DiscriminatorType.STRING)
-public abstract class Pessoa implements Serializable {
-    private static final long serialVersionUID = 1L;
+public class Pessoa {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
@@ -29,8 +22,20 @@ public abstract class Pessoa implements Serializable {
     @NotNull(message = "email é obrigatório")
     protected String email;
 
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "pessoa")
-    protected List<Telefone> telefoneList = new ArrayList<>();
+    @NotNull(message = "tipo pessoa é obrigatório")
+    @Enumerated(EnumType.STRING)
+    protected TipoPessoa tipoPessoa;
+
+//    @OneToMany(cascade = CascadeType.ALL, mappedBy = "pessoa")
+//    protected List<Telefone> telefoneList = new ArrayList<>();
+
+    public TipoPessoa getTipoPessoa() {
+        return tipoPessoa;
+    }
+
+    public void setTipoPessoa(TipoPessoa tipoPessoa) {
+        this.tipoPessoa = tipoPessoa;
+    }
 
     public Long getId() {
         return id;
@@ -56,31 +61,13 @@ public abstract class Pessoa implements Serializable {
         this.email = email;
     }
 
-    public List<Telefone> getTelefones() {
-        return telefoneList;
-    }
-
-    public void setTelefones(List<Telefone> telefones) {
-        this.telefoneList = telefones;
-    }
-
-    public boolean addTelefone(@Valid Telefone telefone) {
-        if (this.telefoneList.stream().anyMatch(tel -> tel.getNumeroCompleto().equals(telefone.getNumeroCompleto())))
-            return false;
-        this.telefoneList.add(telefone);
-        return true;
-    }
-
-    public boolean removeTelefone(Telefone telefone) {
-        Optional<Telefone> telSalvo = this.telefoneList.stream()
-                .filter(tel -> tel.getNumeroCompleto().equals(telefone.getNumeroCompleto()))
-                .findFirst();
-        if (!telSalvo.isPresent()) {
-            this.telefoneList.remove(telSalvo.get());
-            return true;
-        }
-        return false;
-    }
+//    public List<Telefone> getTelefoneList() {
+//        return telefoneList;
+//    }
+//
+//    public void setTelefoneList(List<Telefone> telefoneList) {
+//        this.telefoneList = telefoneList;
+//    }
 
     public <T> void setIfNotNull(final Consumer<T> setter, final T value) {
         if (value != null) {
