@@ -2,7 +2,7 @@ package br.uff.tecnomarias.domain.enums;
 
 public enum PorteEmpresa {
     MICROEMPRESA,
-    EMPRESA_PEQUENO_PORTE,
+    EMRPESA_PEQUENO_PORTE,
     EMPRESA_MEDIO_PORTE,
     GRANDE_EMPRESA
 }
