@@ -1,6 +1,5 @@
 package br.uff.tecnomarias.domain.enums;
 
 public enum TipoPessoa {
-
-    PESSOA_FISICA, PESSOA_JURIDICA;
+    PF, PJ;
 }
