@@ -1,39 +1,49 @@
 package br.uff.tecnomarias.service;
 
-import br.uff.tecnomarias.domain.dao.PessoaFisicaDAO;
+import br.uff.tecnomarias.domain.entity.Links;
 import br.uff.tecnomarias.domain.entity.PessoaFisica;
+import br.uff.tecnomarias.domain.repository.PessoaFisicaRepository;
+import br.uff.tecnomarias.service.exception.EntidadeNaoEncontradaException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
-import javax.annotation.ManagedBean;
-import javax.inject.Inject;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
-import javax.ws.rs.WebApplicationException;
 import java.util.List;
 
-@ManagedBean
+@Service
 public class PessoaFisicaService {
 
-    @Inject
-    PessoaFisicaDAO pfDAO;
+    @Autowired
+    PessoaFisicaRepository pfRepository;
 
     @Transactional
     public PessoaFisica salvar(@Valid PessoaFisica pf) {
-        return pfDAO.salvar(pf);
+        return pfRepository.save(pf);
     }
 
     @Transactional
-    public PessoaFisica alterar(Long id, @Valid PessoaFisica pfAlterada) {
-        PessoaFisica pfSalva = pfDAO.buscarPorIdOptional(id)
-                .orElseThrow(() -> new WebApplicationException("Pessoa nao encontrada", 404));
-        return pfSalva.atualizarDados(pfAlterada);
+    public PessoaFisica alterar(Long id, @Valid Links links) {
+        PessoaFisica pfSalva = pfRepository.findById(id)
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Pessoa nao encontrada"));
+        pfSalva.getLinks().atualizarLinks(links);
+        return pfRepository.save(pfSalva);
     }
 
     public PessoaFisica buscarPorId(Long id) {
-        return pfDAO.buscarPorIdOptional(id)
-                .orElseThrow(() -> new WebApplicationException("Pessoa não encontrada", 404));
+        return pfRepository.findById(id)
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Pessoa não encontrada"));
     }
 
     public List<PessoaFisica> buscarTodas() {
-        return pfDAO.buscarTodas();
+        return pfRepository.findAll();
     }
+
+    @Transactional
+    public void remover(final Long id) {
+        PessoaFisica pf = pfRepository.findById(id)
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Vaga nao encontrada"));
+        pfRepository.delete(pf);
+    }
+
 }
